Use big integers to match Laravel

+++ database/migrations/2016_05_18_000000_teamwork_setup_tables.php
@@ -13,19 +13,19 @@ class TeamworkSetupTables extends Migration
     public function up()
     {
         Schema::table(\Config::get('teamwork.users_table'), function (Blueprint $table) {
-            $table->integer('current_team_id')->unsigned()->nullable();
+            $table->bigInteger('current_team_id')->unsigned()->nullable();
         });
 
         Schema::create(\Config::get('teamwork.teams_table'), function (Blueprint $table) {
-            $table->increments('id')->unsigned();
-            $table->integer('owner_id')->unsigned()->nullable();
+            $table->bigIncrements('id')->unsigned();
+            $table->bigInteger('owner_id')->unsigned()->nullable();
             $table->string('name');
             $table->timestamps();
         });
 
         Schema::create(\Config::get('teamwork.team_user_table'), function (Blueprint $table) {
             $table->bigInteger('user_id')->unsigned();
-            $table->integer('team_id')->unsigned();
+            $table->bigInteger('team_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('user_id')
@@ -41,9 +41,9 @@ class TeamworkSetupTables extends Migration
         });
 
         Schema::create(\Config::get('teamwork.team_invites_table'), function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->bigInteger('user_id')->unsigned();
-            $table->integer('team_id')->unsigned();
+            $table->bigInteger('team_id')->unsigned();
             $table->enum('type', ['invite', 'request']);
             $table->string('email');
             $table->string('accept_token');
